+++ flamerelay/static/js/api.ts
@@ -1,4 +1,4 @@
-function getCsrfToken(): string {
+export function getCsrfToken(): string {
   const match = document.cookie.match(/(?:^|;)\s*csrftoken=([^;]+)/);
   return match ? decodeURIComponent(match[1]) : '';
 }
